Correspondência entre campos e colunas

Cria um vetor de caracteres que especifica as colunas que representam cada campo de endereço em um dataframe.

correspondencia_campos(
  tipo_de_logradouro = NULL,
  logradouro = NULL,
  numero = NULL,
  complemento = NULL,
  cep = NULL,
  bairro = NULL,
  municipio = NULL,
  estado = NULL
)

Arguments

tipo_de_logradouro, logradouro, numero, complemento, cep, bairro, municipio, estado

Uma string. O nome da coluna que representa o respectivo campo de endereço no dataframe. Pode ser NULL, no caso do campo não estar listado. Ao menos um dos campos deve receber um valor não nulo.

Value

Um vetor nomeado de caracteres, em que os nomes representam os campos do endereço e os valores as colunas que os descrevem no dataframe.

Examples

correspondencia_campos(
  tipo_de_logradouro = "tipo_de_logradouro",
  logradouro = "logradouro",
  numero = "nroLogradouro",
  complemento = "complemento",
  cep = "cep",
  bairro = "bairro",
  municipio = "codmun_dom",
  estado = "uf_dom"
)
#>   tipo_de_logradouro           logradouro               numero 
#> "tipo_de_logradouro"         "logradouro"      "nroLogradouro" 
#>          complemento                  cep               bairro 
#>        "complemento"                "cep"             "bairro" 
#>            municipio               estado 
#>         "codmun_dom"             "uf_dom"